\name{gwselect-package}
\alias{gwselect-package}
\alias{gwselect}
\docType{package}
\title{
Variable selection for a geographically weighted regression model, using the LASSO
}
\description{
Variable selection for a geographically weighted regression model, using the LASSO
}
\details{
\tabular{ll}{
Package: \tab gwselect\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-08-24\cr
License: \tab GPL-3\cr
}
\author{
Wesley Brooks
Maintainer: Wesley Brooks <wesley@somesquares.com>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
package
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
